# services/secret-service/doc/openapi.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Secrets Service",
    "description" : "Secrets service is used in the Open Integration Hub for managing secrets used in integration flows.",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "0.1.0"
  },
  "servers" : [ {
    "url" : "/api/v1"
  } ],
...
  "security": [
    {
      "BearerToken": []
    }
  ],
  "paths" : {
    "/auth-clients/" : {
      "get" : {
        "tags" : [ "Auth clients" ],
        "summary" : "Retrieve all clients",
        "description" : "Retrieve all clients created by the current user.",
        "operationId" : "getClients",
        "parameters": [
          {
            "$ref": "#/components/parameters/pageSize"
          },
          {
            "$ref": "#/components/parameters/pageNumber"
          }
        ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AuthClientEntry"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/Meta"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/500ServerError"
          }
        }
      },
      "post" : {
        "tags" : [ "Auth clients" ],
        "summary" : "Create a client",
        "description" : "Create an client",
        "operationId" : "createClient",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MutableAuthClient"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AuthClient"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/401Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/500ServerError"
          }
        }
      }
    },
    "/auth-clients/{clientId}" : {
      "get" : {
        "tags" : [ "Auth clients" ],
        "summary" : "Find client by ID",
        "description" : "Returns a client with given ID",
        "operationId" : "getClientById",
        "parameters" : [ {
          "name" : "clientId",
          "in" : "path",
          "description" : "ID of the client to return",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          },
          "example": "61719514a477890012b804b9"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "properties" : {
                    "data" : {
                      "$ref" : "#/components/schemas/MutableAuthClient"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401Unauthorized"
          }
        }
      },
      "patch" : {
        "tags" : [ "Auth clients" ],
        "summary" : "Update a client",
        "description" : "Update a client",
        "operationId" : "updateClient",
        "parameters" : [ {
          "name" : "clientId",
          "in" : "path",
          "description" : "ID of the client to update",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MutableAuthClient"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "properties" : {
                    "data" : {
                      "$ref" : "#/components/schemas/MutableAuthClient"
                    }
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Client with given ID not found"
          },
          "405" : {
            "description" : "Invalid input"
          }
        }
      }
    },
    "/auth-clients/{clientId}/start-flow" : {
      "post" : {
        "tags" : [ "Auth clients" ],
        "summary" : "Start an authorization flow",
        "description" : "Can be done by any user",
        "operationId" : "startPlatformAuthFlow",
        "parameters" : [ {
          "name" : "clientId",
          "in" : "path",
          "description" : "ID of the client to return",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          },
          "example": "6171ad81a477890012b804c0"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AuthFlow"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "properties" : {
                    "data" : {
                      "properties" : {
                        "authUrl" : {
                          "type" : "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401Unauthorized"
          },
          "405": {
            "$ref": "#/components/responses/405MethodNotAllowed"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Owner" : {
        "required" : [ "id", "type" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "example": "617082258803c70031f85b26"
          },
          "type" : {
            "type" : "string",
            "description" : "Type of Owner entity, e.g. GROUP, TENANT, USER, etc.",
            "example": "USER"
          }
        }
      },
      "MutableAuthClient" : {
        "required" : [ "name", "owners", "type" ],
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "owners" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Owner"
            }
          },
          "type" : {
            "type" : "string",
            "enum" : [ "OA1_TWO_LEGGED", "OA1_THREE_LEGGED", "OA2_AUTHORIZATION_CODE", "SESSION_AUTH" ]
          },
          "preprocessor": {
            "type": "string"
          },
          "tenant": {
            "type": "string"
          }
        },
        "oneOf" : [ {
          "$ref": "#/components/schemas/MutableOA1TwoLeggedClient"
        }, {
          "$ref": "#/components/schemas/MutableOA1ThreeLeggedClient"
        }, {
          "$ref": "#/components/schemas/MutableOA2AuthorizationCodeClient"
        }, {
          "$ref": "#/components/schemas/MutableSessionAuthClient"
        } ],
        "discriminator": {
          "propertyName": "type",
          "mapping": {
            "OA1_TWO_LEGGED": "#/components/schemas/MutableOA1TwoLeggedClient",
            "OA1_THREE_LEGGED": "#/components/schemas/MutableOA1ThreeLeggedClient",
            "OA2_AUTHORIZATION_CODE": "#/components/schemas/MutableOA2AuthorizationCodeClient",
            "SESSION_AUTH": "#/components/schemas/MutableSessionAuthClient"
          }
        }
      },
      "MutableOA1TwoLeggedClient" : {
        "required" : [ "consumerKey", "consumerSecret", "nonce", "signature", "signatureMethod" ],
        "properties" : {
          "consumerKey" : {
            "type" : "string"
          },
          "consumerSecret" : {
            "type" : "string"
          },
          "nonce" : {
            "type" : "string"
          },
          "signature" : {
            "type" : "string"
          },
          "signatureMethod" : {
            "type" : "string"
          },
          "verifier" : {
            "type" : "string"
          }
        }
      },
      "MutableOA1ThreeLeggedClient": {
        "required": ["key", "secret", "redirectUri", "access"],
        "properties": {
          "appName": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "secret": {
            "type": "string"
          },
          "nonce" : {
            "type" : "string"
          },
          "signature" : {
            "type" : "string"
          },
          "signatureMethod" : {
            "type" : "string"
          },
          "endpoints": {
            "type": "object",
            "required": ["request", "authorize", "access"],
            "properties": {
              "request": {
                "type": "string"
              },
              "authorize": {
                "type": "string"
              },
              "access": {
                "type": "string"
              }
            }
          },
          "redirectUri": {
            "type": "string"
          }
        }
      },
      "MutableOA2AuthorizationCodeClient" : {
        "required" : [ "clientId", "clientSecret", "redirectUri", "endpoints"],
        "properties" : {
          "clientId" : {
            "type" : "string",
            "example": "292085223830.apps.googleusercontent.com"
          },
          "clientSecret" : {
            "type" : "string",
            "example": "F4B82757C43741436B195B28AF4AF"
          },
          "redirectUri" : {
            "type" : "string",
            "example": "http://localhost:9001/api/v1/callback"
          },
          "endpoints" : {
            "$ref" : "#/components/schemas/MutableOA2AuthorizationCodeClient_endpoints"
          },
          "predefinedScope": {
            "type": "string"
          }
        }
      },
      "MutableOA2AuthorizationCodeClient_endpoints" : {
        "type" : "object",
        "required": ["auth", "token"],
        "properties" : {
          "auth" : {
            "type" : "string"
          },
          "token" : {
            "type" : "string"
          },
          "userinfo" : {
            "type" : "string"
          },
          "revocation": {
            "type": "string"
          },
          "endSession": {
            "type": "string"
          }
        }
      },
      "MutableSessionAuthClient": {
        "required": ["fields", "endpoints"],
        "properties": {
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionField"
            }
          },
          "tokenPath": {
            "type": "string"
          },
          "endpoints": {
            "type": "object",
            "required": ["auth"],
            "properties": {
              "auth": {
                "$ref": "#/components/schemas/RequestConfig"
              },
              "verification": {
                "$ref": "#/components/schemas/RequestConfig"
              }
            }
          }
        }
      },
      "AuthClient" : {
        "oneOf" : [ {
          "$ref" : "#/components/schemas/OA1TwoLeggedClient"
        }, {
          "$ref": "#/components/schemas/OA1ThreeLeggedClient"
        }, {
          "$ref" : "#/components/schemas/OA2AuthorizationCodeClient"
        }, {
          "$ref": "#/components/schemas/SessionAuthClient"
        } ],
        "discriminator": {
          "propertyName": "type",
          "mapping": {
            "OA1_TWO_LEGGED": "#/components/schemas/MutableOA1TwoLeggedClient",
            "OA1_THREE_LEGGED": "#/components/schemas/MutableOA1ThreeLeggedClient",
            "OA2_AUTHORIZATION_CODE": "#/components/schemas/MutableOA2AuthorizationCodeClient",
            "SESSION_AUTH": "#/components/schemas/MutableSessionAuthClient"
          }
        }
      },
      "OA1TwoLeggedClient" : {
        "allOf" : [{
          "required" : [ "createdAt", "_id" ],
          "type" : "object",
          "properties" : {
            "_id" : {
              "type" : "string"
            },
            "createdAt" : {
              "type" : "string",
              "description" : "Client creation time",
              "format" : "date-time"
            },
            "updatedAt" : {
              "type" : "string",
              "description" : "Client update time",
              "format" : "date-time"
            }
          }
        }, {
          "$ref" : "#/components/schemas/MutableOA1TwoLeggedClient"
        } ]
      },
      "OA1ThreeLeggedClient": {
        "allOf" : [ {
          "$ref" : "#/components/schemas/MutableOA1ThreeLeggedClient"
        }, {
          "required" : [ "createdAt", "_id" ],
          "type" : "object",
          "properties" : {
            "_id" : {
              "type" : "string"
            },
            "createdAt" : {
              "type" : "string",
              "description" : "Client creation time",
              "format" : "date-time"
            },
            "updatedAt" : {
              "type" : "string",
              "description" : "Client update time",
              "format" : "date-time"
            }
          }
        } ]
      },
      "OA2AuthorizationCodeClient" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/MutableOA2AuthorizationCodeClient"
        }, {
          "required" : [ "createdAt", "_id" ],
          "type" : "object",
          "properties" : {
            "_id" : {
              "type" : "string",
              "example": ""
            },
            "createdAt" : {
              "type" : "string",
              "description" : "Client creation time",
              "format" : "date-time"
            },
            "updatedAt" : {
              "type" : "string",
              "description" : "Client update time",
              "format" : "date-time"
            }
          }
        } ]
      },
      "SessionAuthClient": {
        "allOf": [ {
          "$ref": "#/components/schemas/MutableSessionAuthClient"
        }, {
          "required" : [ "createdAt", "_id" ],
          "type" : "object",
          "properties" : {
            "_id" : {
              "type" : "string"
            },
            "createdAt" : {
              "type" : "string",
              "description" : "Client creation time",
              "format" : "date-time"
            },
            "updatedAt" : {
              "type" : "string",
              "description" : "Client update time",
              "format" : "date-time"
            }
          }
        }]
      },
      "SessionField": {
        "type": "object",
        "required": ["key"],
        "properties": {
          "key": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "required": {
            "type": "boolean"
          }
        }
      },
      "RequestField": {
        "type": "object",
        "required": ["key"],
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      },
      "RequestConfig": {
        "type": "object",
        "required": ["requestFields", "authType", "url"],
        "properties": {
          "requestFields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RequestField"
            }
          },
          "label": {
            "type": "string"
          },
          "authType": {
            "type": "string",
            "enum": ["HEADER_AUTH", "BODY_AUTH", "PARAMS_AUTH"]
          },
          "url": {
            "type": "string"
          }
        }
      },
      "AuthClientEntry": {
        "type": "object",
        "required": ["_id", "name", "type"],
        "properties": {
          "_id": {
            "type": "string",
            "example": "616a3783d25a1e002b98f6e0"
          },
          "name": {
            "type": "string",
            "example": "My Auth Client"
          },
          "type": {
            "type": "string",
            "example": "OA2_AUTHORIZATION_CODE"
          }
        }
      },
...
      "AuthFlow" : {
        "type" : "object",
        "properties" : {
          "scope" : {
            "type" : "string"
          },
          "secretName" : {
            "type" : "string"
          },
          "successUrl" : {
            "type" : "string"
          }
        }
      },
      "Meta" : {
        "type" : "object",
        "properties" : {
          "page" : {
            "type" : "integer",
            "example": 1
          },
          "perPage" : {
            "type" : "integer",
            "example": 10
          },
          "total" : {
            "type" : "integer",
            "example": 15
          },
          "totalPages" : {
            "type" : "integer",
            "example": 2
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      }
    },
    "parameters": {
      "pageSize": {
        "name": "page[size]",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 10
        },
        "description": "The number of results to return per page",
        "required": false,
        "style": "deepObject"
      },
      "pageNumber": {
        "name": "page[number]",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 1
        },
        "description": "The page number to return (not zero indexed)",
        "required": false
      }
    },
    "responses": {
      "400BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "401Unauthorized": {
        "description": "Unauthorized operation",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "405MethodNotAllowed": {
        "description": "Method not allowed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "500ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "securitySchemes" : {
      "BearerToken": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}